iio: ad9467: ad9680: Add helper function for writing channel registers
Some registers in the AD9680 register map are channel specific and can take
different values for each channel. Which channels are accessed when writing
to the register is selected by the DEVICE_INDEX register.

Add a helper function for writing channel specific registers that
implements the pattern of selecting a channel through the DEVICE_INDEX
register, writes the target register and then selects all channels again.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>