iio: ad9517: Fix clock source selection
The AD9517 (and similar) allows three options for the clock source to the clock distribution network.
 1) External clock
 2) External clock through VCO divider
 3) Internal VCO through VCO divider

At the moment the second case is not handled correctly since the driver
code assumes that if the VCO divider is not bypassed the VCO is used. But
the VCO divider input might be connected to the external clock.

Handle the case correctly and only set 'uses_vco' to true if the VCO
divider input is the VCO output.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>