iio: ad9517: Allow firmware-less operation
Currently the AD9517 driver requires a firmware file that contains all the
initial register settings. When supporting many boards that use the AD9517
this can get messy quickly as each board needs its own firmware file
compiled into the kernel.

For simple setups that do not use the PLL the driver already provides
enough functionality to program the output dividers and enable the clock
outputs.

In order to allow firmware less operation for such setups make the firmware
file optional. When no firmware file is specified use the power-on reset
register settings. With one minor modification, the power-on reset register
setting have every second output powered up by default. Change this so that
all outputs are powered down by default.

The clock API can be used to set the frequency and enable an output as
necessary.

If required a output frequency can also be specified in the devicetree
using the assigned-clocks bindings.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>