iio: adxl372: Fix revid check
Before the 3rd revision of the chip, the I2C communication was not
working properly if there were any other I2C devices present on the bus.
Howerver, if adxl372 is the only device, then it should still work, this
is why it is enough to just print a warning and continue with the device
probe.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>